Remove lint-markdown and lint-shell workflows, add lint workflow, and update README.md (#10)
This pull request includes changes related to the linting and formatting
of files. The most important changes include the removal of the workflow
for linting markdown files and the addition of a new workflow for
linting files, triggering on pull requests and pushes to the `main`
branch.

Linting and formatting changes:

*
[`.github/workflows/lint-markdown.yml`](diffhunk://#diff-29231bb7ac9ddd9de0d2feec83208bbee635cb8b2ddcf3b821b71bed20b2e3ceL1-L28):
Removed the workflow for linting markdown files, including the trigger
on pull requests and the prettifying job.
*
[`.github/workflows/lint-shell.yml`](diffhunk://#diff-cb01ce593de6f1d1a9c247dde8298db8c7210be7d0654e00597dd13c8e0e565eL1-L21):
Removed the workflow for linting shell scripts and the jobs responsible
for checking the syntax of shell scripts using specific actions.
*
[`.github/workflows/lint.yml`](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2R1-R66):
Added a new workflow for linting files, triggering on pull requests and
pushes to the `main` branch. Added jobs to find changed shell and
markdown files, check shell script syntax, and prettify markdown files.

Other changes:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L3):
Removed the CodeFactor badge from the README.md file.