Add SonarCloud workflow for continuous code quality analysis (#22)
This pull request introduces a new GitHub Actions workflow for SonarQube
analysis. The main changes include setting up the workflow to run on
specific events and configuring the necessary steps for the SonarQube
scan.

### New GitHub Actions workflow for SonarQube analysis:

*
[`.github/workflows/sonarcloud.yml`](diffhunk://#diff-5895707186d21bd3672c2faf7743c1cd2d656de088cdec80dc676ed8e3f22bceR1-R19):
Added a new workflow named `Build` that triggers on `push` to the `main`
branch and on pull request events (`opened`, `synchronize`, `reopened`).
The workflow includes a job named `SonarQube` that runs on
`ubuntu-latest` and performs a SonarQube scan using the
`SonarSource/sonarqube-scan-action`.