Add MegaLinter workflow for automated code quality checks (#23)
This pull request includes several changes to improve the functionality
and maintainability of the codebase, specifically in the CI/CD workflows
and the `check_gmirror.sh` script. The most important changes include
adding a MegaLinter workflow, updating permissions for the SonarCloud
workflow, and refactoring the `check_gmirror.sh` script for better
readability and performance.

### CI/CD Workflow Improvements:
* Added a new MegaLinter workflow in `.github/workflows/megalinter.yml`
to automate code linting and fixing. This workflow runs on pull requests
targeting the `master` and `main` branches and includes configurations
for applying fixes automatically.
* Updated the SonarCloud workflow in `.github/workflows/sonarcloud.yml`
to include necessary permissions for SonarQube to read code, create
issues, and upload results.

### Script Refactoring:
* Refactored the `check_gmirror.sh` script to improve readability by
combining single-line `if` statements with their corresponding `then`
statements. This change simplifies the script and makes it easier to
follow.
[[1]](diffhunk://#diff-6fe28a0e0819eb06e45f4066636701348a5266cd1e86c9bfb8d8e97eb23360d5L8-R27)
[[2]](diffhunk://#diff-6fe28a0e0819eb06e45f4066636701348a5266cd1e86c9bfb8d8e97eb23360d5L50-R51)

---------

Co-authored-by: Klintrup <18490009+Klintrup@users.noreply.github.com>
Co-authored-by: Klintrup <Klintrup@users.noreply.github.com>